Extract common code from tests to helpers
