+++ ydb/core/tx/schemeshard/ut_helpers/ya.make
@@ -28,6 +28,7 @@ PEERDIR(
 SRCS(
     auditlog_helpers.cpp
     auditlog_helpers.h
+    data_erasure_helpers.cpp
     export_reboots_common.cpp
     failing_mtpq.cpp
     helpers.cpp
